+++ src/OrderedList.php
@@ -15,7 +15,7 @@ use function implode;
 use function is_array;
 
 /**
- * @implements MemberList<int, Item|InnerList<int, Item>>
+ * @implements MemberList<int, Item|MemberList<int, Item>>
  */
 final class OrderedList implements MemberList
 {
@@ -27,13 +27,18 @@ final class OrderedList implements MemberList
         $this->members = array_values($members);
     }
 
-    public static function from(InnerList|Item|ByteSequence|Token|bool|int|float|string ...$members): self
+    /**
+     * @param MemberList<int, Item>|Item|ByteSequence|Token|bool|int|float|string ...$members
+     *
+     * @return static
+     */
+    public static function from(MemberList|Item|ByteSequence|Token|bool|int|float|string ...$members): self
     {
         return self::fromList($members);
     }
 
     /**
-     * @param iterable<InnerList|Item|ByteSequence|Token|bool|int|float|string> $members
+     * @param iterable<MemberList<int, Item>|Item|ByteSequence|Token|bool|int|float|string> $members
      */
     public static function fromList(iterable $members = []): self
     {
@@ -102,7 +107,7 @@ final class OrderedList implements MemberList
     }
 
     /**
-     * @return Iterator<int, Item|InnerList<int, Item>>
+     * @return Iterator<int, Item|MemberList<int, Item>>
      */
     public function getIterator(): Iterator
     {
@@ -181,7 +186,7 @@ final class OrderedList implements MemberList
     /**
      * Inserts members at the beginning of the list.
      */
-    public function unshift(StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function unshift(MemberList|StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
     {
         $this->members = [...array_map(self::filterMember(...), array_values($members)), ...$this->members];
 
@@ -191,7 +196,7 @@ final class OrderedList implements MemberList
     /**
      * Inserts members at the end of the list.
      */
-    public function push(StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function push(MemberList|StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
     {
         $this->members = [...$this->members, ...array_map(self::filterMember(...), array_values($members))];
 
@@ -203,7 +208,7 @@ final class OrderedList implements MemberList
      *
      * @throws InvalidOffset If the index does not exist
      */
-    public function insert(int $index, StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
+    public function insert(int $index, MemberList|StructuredField|ByteSequence|Token|bool|int|float|string ...$members): self
     {
         $offset = $this->filterIndex($index);
         match (true) {
@@ -281,12 +286,12 @@ final class OrderedList implements MemberList
 
     /**
      * @param int|null $offset
-     * @param InnerList<int, Item>|Item|ByteSequence|Token|bool|int|float|string $value  the member to add
+     * @param MemberList<int, Item>|Item|ByteSequence|Token|bool|int|float|string $value  the member to add
      *
      * @see ::push
      * @see ::replace
      */
-    public function offsetSet(mixed $offset, $value): void
+    public function offsetSet(mixed $offset, mixed $value): void
     {
         if (null !== $offset) {
             $this->replace($offset, $value);