+++ src/ParametersTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Bakame\Http\StructuredFields;
 
-use TypeError;
 use function iterator_to_array;
 
 /**
@@ -350,8 +349,8 @@ final class ParametersTest extends StructuredFieldTest
     /** @test */
     public function it_throws_if_the_structured_field_is_not_supported(): void
     {
-        $this->expectException(TypeError::class);
+        $this->expectException(InvalidArgument::class);
 
-        Parameters::fromPairs(['foo', InnerList::from(42)]); // @phpstan-ignore-line
+        Parameters::fromPairs([['foo', InnerList::from(42)]]); // @phpstan-ignore-line
     }
 }